Updates to post deletion logic.  Moved the initial status change from adminhandler to post::delete().  In this way, one need only ever call $post->delete() to have that post's status set to "deleted".  Subsequent calls to $post->delete() will purge the post from the database.
Post::delete() also accepts an optional boolean argument.  This 
argument, if true, will immediately delete the post.  The defautl value 
for this option is false.

Implemented plugin hooks for post deletion, as well as changes to a 
post objections member properties.  post_delete actions will be 
triggered when deleting a post; and post_update_* will be triggered when 
making changes to a post object in Post::update().



git-svn-id: https://svn.habariproject.org/habari/trunk/htdocs/system@589 653ae4dd-d31e-0410-96ef-6bf7bf53c507

+++ classes/post.php
@@ -387,7 +387,13 @@ class Post extends QueryRecord
 	{
 		$this->updated = date('Y-m-d h:i:s');
 		if(isset($this->fields['guid'])) unset( $this->newfields['guid'] );
-		//$this->setslug();  // setslug() for an update?  Hmm.  No?
+		// invoke plugins for all fields which have been changed
+		// For example, a plugin action "update_post_status" would be
+		// triggered if the post has a new status value
+		foreach ( $this->newfields as $fieldname => $value )
+		{
+			Plugins::act('update_post_' . $fieldname, $this, $value );
+		}
 		$result = parent::update( DB::table('posts'), array('slug'=>$this->slug) );
 		$this->fields = array_merge($this->fields, $this->newfields);
 		$this->newfields = array();
@@ -398,9 +404,20 @@ class Post extends QueryRecord
 	/**
 	 * function delete
 	 * Deletes an existing post
+	 * @param Boolean whether to delete the post immediately, or set its status to "deleted"
 	 */	 	 	 	 	
-	public function delete()
+	public function delete( $delete_now = FALSE )
 	{
+		if ( ! $delete_now )
+		{
+			$this->status= Post::status('deleted');
+			$this->update();
+			return;
+		}
+
+		// invoke plugins
+		Plugins::act('post_delete', $this);
+
 		// Delete all comments associated with this post
 		if(!empty($this->comments))
 			$this->comments->delete();
